move the class into matched namespace.
